Sort the generated kernel list
The kernel list is normally sorted automatically via `sort -V`, but a
pinned kernel will (likely) not be at the end of the list. Due to fzf
--tac, the item on the end of the list is auto-selected when entering
the kernel screen. This pops the kernel that will be booted to the end
of the list, ensuring that it's automatically chosen for the user.